Refactoring the home page to use Movies Context and Hook
