+++ front/src/components/MovieCard/style.scss
