+++ front/src/views/Home/style.scss
@@ -1,10 +1,6 @@
 .home {
 
-    .home__main {
-        display: flex;
-        flex-wrap: wrap;
-        justify-content: center;
-        column-gap: 55px;
-        row-gap: 45px;
+    .inputText {
+        margin: 35px 0px;
     }
 }
